pyproject.toml
[project]
name = "fastapi-openapi-cli"
version = "0.1.0"
description = "CLI tool to export FastAPI OpenAPI specifications"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Hector Castro", email = "hectcastro@gmail.com" }]
requires-python = ">=3.12"
dependencies = ["fastapi>=0.115.0", "pyyaml>=6.0.2"]

[project.scripts]
fastapi-openapi = "fastapi_openapi_cli.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a25",
    "deptry>=0.21.0",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TC",  # flake8-type-checking
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--cov=fastapi_openapi_cli",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=100",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.deptry]
extend_exclude = ["tests"]
known_first_party = ["fastapi_openapi_cli"]
